requirements.txt
langchain
langchain-community
chromadb
requests
google-generativeai
sentence-transformers
ollama
jupyter
python-dotenv
